tasks/serializers.py
from rest_framework import serializers
from .models import Task

class TaskSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Task
        fields = ['id', 'group', 'title', 'description', 'category', 'urgency', 'due_date']