requirements.txt
alabaster==0.7.12
altgraph==0.17.2
asgiref==3.4.1
auto-py-to-exe==2.18.2
Babel==2.9.1
beautifulsoup4==4.11.1
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.2
colorama==0.4.4
Django==3.2.9
djangorestframework==3.13.1
docutils==0.17.1
Eel==0.12.4
Flask==2.1.1
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
imagesize==1.3.0
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.0.1
MouseInfo==0.1.3
packaging==21.3
Panda3D==1.10.11
panda3d-gltf==0.13
panda3d-simplepbr==0.10
pefile==2021.9.3
Pillow==9.1.0
psycopg2==2.9.3
py2exe==0.11.1.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.10.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
PyMsgBox==1.0.9
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.3
pywin32-ctypes==0.2.0
requests==2.26.0
screeninfo==0.8
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.3.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.7
ursina==4.1.1
Werkzeug==2.1.1
whichcraft==0.6.1
wikipedia==1.4.0
zope.event==4.5.0
zope.interface==5.4.0
